more refactors of headers into submodules
