+++ crates/skytem_logs/src/mbed_motor_control/pid.rs
@@ -1,4 +1,5 @@
 pub mod entry;
-pub mod header_v1;
-pub mod header_v2;
+pub mod header;
 pub mod pidlog;
+
+pub(crate) const UNIQUE_DESCRIPTION: &str = "MBED-MOTOR-CONTROL-PID-LOG-2024";
