+++ crates/skytem_logs/src/mbed_motor_control/pid/header/v2.rs
@@ -15,7 +15,7 @@ use serde::{Deserialize, Serialize};
 use serde_big_array::BigArray;
 
 #[derive(Debug, PartialEq, Deserialize, Serialize, Clone, Copy)]
-pub struct PidLogHeaderV2 {
+pub(crate) struct PidLogHeaderV2 {
     #[serde(with = "BigArray")]
     unique_description: UniqueDescriptionData,
     version: u16,
@@ -102,7 +102,6 @@ impl GitMetadata for PidLogHeaderV2 {
 
 impl MbedMotorControlLogHeader for PidLogHeaderV2 {
     const VERSION: u16 = 2;
-    const UNIQUE_DESCRIPTION: &'static str = "MBED-MOTOR-CONTROL-PID-LOG-2024";
     /// Size of the header type in bytes if represented in raw binary
     const RAW_SIZE: usize = SIZEOF_UNIQ_DESC
         + SIZEOF_PROJECT_VERSION
@@ -194,7 +193,7 @@ mod tests {
         eprintln!("{pid_log_header}");
         assert_eq!(
             pid_log_header.unique_description(),
-            PidLogHeaderV2::UNIQUE_DESCRIPTION
+            crate::mbed_motor_control::pid::UNIQUE_DESCRIPTION
         );
         assert_eq!(pid_log_header.version, 2);
         assert_eq!(pid_log_header.project_version().unwrap(), "2.0.2");
