+++ crates/skytem_logs/src/mbed_motor_control/status.rs
@@ -1,3 +1,5 @@
-pub mod entry;
+mod entry;
 mod header;
 pub mod statuslog;
+
+pub(crate) const UNIQUE_DESCRIPTION: &str = "MBED-MOTOR-CONTROL-STATUS-LOG-2024";
